Resolves LightBulbVisualizer is Obsolete (#13889)
